ANDREW_WEATHER_PROJECT/PythonProject/requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decompyle3==3.9.2
Django==5.0.6
django-debug-toolbar==4.4.6
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.0
opencv-python==4.10.0.84
pandas==2.2.2
pillow==11.0.0
pyTelegramBotAPI==4.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-git==2018.2.1
pytz==2024.1
requests==2.32.3
Send2Trash==1.8.3
six==1.16.0
soupsieve==2.5
spark-parser==1.8.9
SQLAlchemy==2.0.35
sqlparse==0.5.0
telebot==0.0.5
tk==0.1.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
waitress==3.0.0
Werkzeug==3.0.4
xdis==6.1.1
XlsxWriter==3.2.0
